Fix apple_static_framework_import duplication
Previously we passed every dependent framework binary on the command
line to the link invocation. This resulted in duplicate symbols in the
case that multiple dependencies depended on the same
apple_static_framework_import. This now treats static frameworks just
like dynamic frameworks and uses `-F` + `-framework` to discover these.
Allowing the linker to dedup them.