Drop partial swiftmodule from other_outputs
Previously we required this file to be produced by adding it to the
outputs of compile actions. Starting with Swift 5.6 in incremental mode
the driver no longer produces these files, and instead separates
compilation from a single module creation step, instead of producing
partial modules and merging them. This leads to this file never being
created, and a build failure. This file is never consumed elsewhere, so
we don't need to ensure it's created. We still leave it in the output
file map so that in the case it is created it's still a validate
filepath.

https://github.com/apple/swift-driver/commit/e0c3e5c740e68202e0a7777d3f6af795d9150296

Fixes https://github.com/bazelbuild/rules_swift/issues/775