Add private feature noting emit-module-separately behavior
With Swift 5.6 the swift driver now emits modules in a separate step,
which skips the partial module merging step that happened before. This
means partial swiftmodule files are not emitted. This reflects that in
our logic for determining this. It's also enabled by default for WMO
builds, but that shouldn't matter for now either since we already
assumed that.

Technically this isn't a perfect solution, this ignores the fact that
users could pass -no-emit-modules-separately, or the WMO version of
these flags but this doesn't seem to be an issue.

https://github.com/apple/swift-driver/commit/e0c3e5c740e68202e0a7777d3f6af795d9150296

Fixes https://github.com/bazelbuild/rules_swift/issues/775