Fix incremental compilation with Swift 5.6
Previously we were trying to copy swiftmodule files based on their
target's name to the incremental source directory. The swiftmodule files
are actually based on the module name instead, which meant this copy was
silently failing. With Swift 5.6 and the separate job for emitting
swiftmodules without first generating partial swiftmodules resulted in
these files not being created when only changes that didn't require a
new swiftmodule were made. This resulted in a bazel failure.

This change fixes this by correctly using the swiftmodule path which is
based on the module name, instead of deriving the path from the output
file map, which is based on the target name. This also adds an error in
the case incremental inputs don't exist when the compile is otherwise
successful, because this means the next job will not have these
potentially required files.