common: remove `file_system` library
Remove the custom `file_system` library in favour of `std::filesystem`.
This allows us to avoid having to re-implement the functionality on all
the platforms.  This will also enable us to support Windows eventually
without having to do all the file system handling (which requires a
significant amount of UTF-8 to UCS-2 conversions) and trying to emulate
the semantics of the functions.