common: port `temp_file` to C++17 and Windows
Use C++17 more heavily here to avoid the use of `fts.h` making the code
easier to port.  Reduce the string manipulation by using
`std::filesystem::path` and use that to additionally locate the
temporary file location.  Emulate the POSIX semantics for `mkdtemp` and
`mkstemp` as much as possible to provide equivalent behaviour on
Windows.