Use `absl::Status` to report errors inside the worker.
Worker-specific issues, like failures when preparing the incremental storage area or copying the outputs back to the Bazel output root, are now reported to the worker response stream instead of stderr. The former is shown to the user (and thus more discoverable); the latter only goes to worker logs that are not easy for end users to find.

This change also makes it recoverable if the worker is unable to create the directory structure for the incremental storage area; it will emit a warning and then fall back to regular non-incremental compilation.

PiperOrigin-RevId: 442814455