ci: add an initial configuration for Windows builds
Add a build-only partial CI coverage for Windows.  This should ensure
that we do not regress on the C++ portions of the Windows support.
Having this coverage should enable us to begin focusing on the Starlark
rules support for Windows.

Update the .bazelrc to use a feature for the spelling of `-Werror` to be
compatible with different compilers.

Replace the C++20 designated initializers with C++17 compatible code.