Un-break and improve incremental compilation when using the `worker` strategy.
Incremental compilation does **not** yet work when using the `dynamic` strategy,
due to sandboxing; however, compilation will fall back to a non-incremental mode and still succeed. This will be addressed later.

When we switched to `string_view`s in many places, we misused them by calling `data()` and passing those `char` pointers to C functions. There is no guarantee however that a `string_view` is null-terminated; especially in the case of computing the parent directory of a path, they're pretty much guaranteed **not** to be.

Other fixes in this change include:

*   In incremental mode, emitting the `.swiftmodule` file
    (and its `doc` and `sourceinfo` friends) to the
    incremental area, so that the compiler can schedule or
    omit merge-module jobs correctly (instead of always
    scheduling them even when the module wouldn't change).
*   Stop declaring partial `.swiftmodule` files at analysis
    time, instead only adding them to the output file map
    when it is rewritten for incremental mode.

PiperOrigin-RevId: 443107738