swift_toolchain: update the toolchain rules for Windows
Introduce a `SDKROOT` attr which is configured initially via the
environment.  This can then be passed along to the compiler
appropriately (`-sdk SDKROOT`).  This attribute also provides a
number of subsequent other values:
- `../../../Info.plist` contains version information for XCTest
- `../../Library/XCTest-$(VERSION)/usr/lib/swift/{os}` XCTest modules
- `../../Library/XCTest-$(VERSION)/usr/lib/swift/{os}/{arch}` XCTest libraries
- `../../Library/XCTest-$(VERSION)/usr/bin{32,64,32a,64a}` XCTest runtime

It can also be used for additional information:
- `SDKSettings.plist` contains additional flags that we should pass to
  the compiler

Add the new `xctest_version` attribute to allow the toolchain to specify
the XCTest version information which is required when building with the
Windows SDK.

Introduce features for `swift-autolink-extract` as these are platform
level configuration rather than build host and are not needed when
targeting PE/COFF or MachO.

Add a `linkstatic` attribute on `swift_library` to enable static and
dynamic linking for Windows.