Add `compilation_context` to `SwiftInfo`
This adds a new struct to `SwiftInfo` that exposes the information used during compilation of a Swift target. The main use case is aspects, in particular ones for IDE integration. The attributes exposed mirror some of those found in `CcInfo.compliation_context` and `ObjcProvider`.

The initial attributes are ones that I personally need right now, but we can add more later as they are identified.