swift: introduce a feature to control debug info on Windows
Windows poorly supports DWARF (or rather DWARF poorly supports Windows)
and the swift driver poorly supports CodeView.  However, both are useful
in different scenarios.  Add a feature to use codeview, and enable it by
default.  The default is important as DWARF cannot supported by link
which is picked up as the default linker.