When calculating the target triple for 'arm64-apple-ios<version>-simulator' targets, correctly specify the minimum supported OS.
Precompiled pcm files will be generated with the target triple, however the minimum supported version for arm64 on an ios-simulator is 14.0. Any minimum OS under 14 will be raised to 14 by the swift driver and clang tools. This logic is not applied to the AST serialized into pcm object files. To ensure the AST is correctly specified, adjust the target triple to configure the lower bounds correctly.

PiperOrigin-RevId: 458248769