Add index-while-building support (back) to the Swift build rules.
When the `swift.index_while_building` feature is set, targets will emit an indexstore directory to the "indexstore" output group.

Unlike the original implementation, this one is focused on the remote indexing use case, so we don't support the old behavior meant for local builds of scanning `--swiftcopt` for a manually-specified path that would be the destination for all index data. Users who want to do that should just pass _all_ of the indexing flags to `--swiftcopt` instead of using this feature.

PiperOrigin-RevId: 459613898