Factor out `-index-ignore-system-modules` into a separate feature
The feature has opposite polarity to the compiler flag, as we want it to be off by default and only enabled on certain targets.

PiperOrigin-RevId: 467693984
(cherry picked from commit 14e066f3f0d9c5c2fe6951000cb6fbcccf896497)

---

Cherry-pick notes: We already had this flag, just named the opposite way. Not worth the breaking change to rename it. Though we are now enabling the flag by default.

Signed-off-by: Brentley Jones <github@brentleyjones.com>