Add basic sanitizer support for swift_test
rules_apple handles bundling sanitizers with test bundles that it
creates, swift_test does not handle this as it creates a very minimal
xctest bundle (ideally it wouldn't have to create that bundle at all and
you could run tests directly). Because of this we don't have the
sanitizer libs easily extracted into the bundle. Sanitizers require us
to load them when launching the binary with DYLD_INSERT_LIBRARIES, to
know which to load we can inspect the binary for these. Clang inserts an
absolute rpath to the Xcode directory containing the libraries, as well
as the library loads for each we depend on.