Implement layering checks for Swift dependencies.
This works by having the build rules generate a list of the module names of direct dependencies and passing that to the worker, which spawns a second `swiftc` invocation with `-emit-imported-modules` to list the modules actually imported in source. The second compilation is a parse-only mode, so it should be fast (especially compared to making it a second, independent Bazel action).

The feature is not yet enabled by default, as it needs to be tested against real code to see if any edge cases are unhandled.

PiperOrigin-RevId: 468305263