Rearchitect command line argument handling for Swift worker sub-jobs, like layering checking.
The layering check sub-job needs to be able to remove flags that aren't relevant from the compilation. To do this, we have to retain the processed command line argument vector (expanding response files unconditionally) and only write them as response files just before the job executes.

Also, change use of `kBoldWhite` to just `kBold` to remove my white-on-black terminal bias.

PiperOrigin-RevId: 469446670