Add a `discover_tests` attribute to `swift_test`.
This attribute, which is `True` by default, controls whether `XCTest`-style tests are automatically discovered (either using the Objective-C runtime or manual discovery via symbol graphs). If set to `False`, the test is assumed to provide its own `main`.

This will subsume the behavior of the `swift.bundled_xctests` feature, which will be removed in the future.

PiperOrigin-RevId: 471589475