Apply default cpu based on target Apple platform
If `--apple_platform_type` is passed but `--cpu` or
`--{ios_multi|watchos|tvos|macos}_cpus` is not passed, apply a transition
to change the `--cpu` flag to a default `cpu` based on the target
platform.

The transitions.bzl was copied from
https://github.com/bazelbuild/bazel/blob/5604779c98edee9bedbe668ca72f2aa1ffb4cf2e/src/main/starlark/builtins_bzl/common/objc/transitions.bzl#L1
with the `//command_line_option:incompatible_enable_apple_toolchain_resolution`
flag removed to continue supporting Bazel 5.x.