Make DEVELOPER_DIR required on Apple platforms (#903)
This is a speculative fix for an issue we saw where swift.remap_xcode
seemed to not take effect. My only theory is that DEVELOPER_DIR was not
set, so the value was not replaced correctly in the invocation, leading
to absolute paths leaking into the binary. With this change this case
would now fail, which make make it easier to debug why bazel didn't set
this required environment variable. This also ignores these variables on
other platforms, where I don't think we'd ever want them to be consulted
even if they were manually set by someone.