Add new swift.modular_indexing feature to index explicit PCMs
This allows us to index explicit PCMs once they are built and then
never again. Currently we only do this for system SDKs as we assume
that we already have source code available for other SDKs, but this
can be expanded in the future.

In addition, make swift.file_prefix_map also remap DEVELOPER_DIR.

PiperOrigin-RevId: 478594522