Add the `module_mapping` flag to the Swift rules.
The value of this flag should be the label a `swift_module_mapping` target, which provides the aliases to use during the build.

At this time, module aliases are only being supported as a globally-applied build flag. This can be revisited in the future if aliases are needed for individual subgraphs, but this is much more complex to represent in BUILD.

This change only adds the flag and hooks it into the toolchain; the compilation changes will be added in a subsequent change.

PiperOrigin-RevId: 486659250