Add back uses of `-disable-autolink-framework` for static frameworks (#941)
This was removed a while back because it appeared to be unused. Since
then this rules_apple change:

https://github.com/bazelbuild/rules_apple/commit/8d841342c238457896cd7596cc29b2d06c9a75f0
causes it to be slightly necessary again. The situation is that we
compile Swift files with the framework search paths, but then we now
link by passing the static framework binary directly to the linker, as
opposed to using `-F` / `-framework`, which means any `LC_LINKER_OPTION`
`-framework` uses pointing to these static frameworks are invalid.
Invalid `LC_LINKER_OPTION`s are ignored by ld64 _unless_ the entire link
fails, but lld and mold do not implement this workaround, so it fails
the link. This also potentially helps with cases of including undeclared
dependencies. Ideally we could disable autolinking all together to
ensure dependencies were strict, but that requires us to manually link
the various Swift support dylibs.

This discovers framework names based on their search paths similar to
how bazel did before the rules_apple change
https://github.com/bazelbuild/bazel/blob/8bd2d97fd143b4b8a6fc96ce300820e6e422c03a/src/main/java/com/google/devtools/build/lib/rules/objc/ObjcProvider.java#L689-L697

We cannot revert the rules_apple change since theoretically that field
or provider will be removed entirely in the future.

This reverts commit b37a7173dab6a33fea89d79e60592594927b7931.