Fix sanitizers on Linux
When a feature name doesn't start with `swift.` we check if it's enabled
by checking against the C++ features, not against all features.
Previously the Linux CC toolchain didn't support sanitizers so these
just never worked on Linux and the features were silently ignored from
Swift. Today these features were added to the Linux CC toolchain
https://github.com/bazelbuild/bazel/commit/abae5ca3e8142f93cf0c2597e3410ed955c4dd59
which broke Linux CI because ubsan doesn't work with Swift on Linux and
asan fails because of a memory leak in Swift