Revert "Add support for imported_library when linking (#917)"
With recent rules_apple updates, libraries are added both to CcInfo and
ObjcProvider's imported_library, which results in duplicate symbols.
Removing this and only relying on CcInfo for this is the path moving
forward. Theoretically a custom rule that produced objc info with
imported_librarys would be broken by this, but they will have to migrate
either way as this provider goes away for linking.

This reverts commit b83db34c861c1dec08c2907afd2bb9850ac88d62.