Remove `swift_module_alias`
Was originally intended to be like `alias` and just used for migration,
but is easily misused to merge/reexport targets, which in turns causes
problems for indexing, automation, etc. that try to map symbols back to
targets.

Since it generates a new files, if dependent on multiple things to merge,
that ends up being a choke point that slows downs builds when someone
really only needs a subset of the dependencies.

So just remove it so it isn't something folks will try to reach for with
many down sides.

If someone really needs the rexport steps for a migration, they can manually
make a targets/srcs to do it, and hopefully thus be aware of the costs so
there will migrate off it as soon as possible.

PiperOrigin-RevId: 528836559