Add support for C++ interop.
This change adds the `cxx_interop` attribute to the compiling rules (`swift_{binary,compiler_plugin,library,test}`). These rules now have a configuration transition that updates a Boolean build setting based on the value of that attribute, and the compiling APIs automatically apply this value when configuring actions.

I debated for a while whether this should be just an attribute on `swift_interop_hint` (in other words, Obj-C/C++ libraries would also dictate which mode they supported). For most cases that would work, but system framework dependencies wouldn't work in this mode; we don't want to have completely independent targets for "C++ mode" vs. "not C++ mode". And the C++ interop mode of an explicit module has to match that of the Swift module that's importing it (transitively, except when cut off by implementation-only imports).

This leaves open the possibility for a module to be built in both configurations through different paths in the same build graph. That's already not valid in Swift though, so we're not creating any new problems by doing this.

PiperOrigin-RevId: 542899538