Restrict `cxx_interop` to binary rules, making it affect the entire dependency tree.
Allowing `cxx_interop` to be set on a per-library basis is _extremely_ prone to duplicate linker inputs, because it's possible to have the same underlying `cc_library` reachable from both a library with C++ interop enabled and one without it, and these are built in different configurations and treated separately. There's no obvious workaround for this that doesn't require some hackery. Instead, we'll allow top-level targets to opt-in for their entire dependency tree and aim to enable it globally (without opt-out) in the future when it's stable enough.

PiperOrigin-RevId: 552580394