Change`swift_library` `alwayslink` default to True.
This also changes the default for the `swift_common.create_linking_context_from_compilation_outputs` api.

SwiftPM works by linking all the .o files, and how .swift sources map
to .o files can be dependent on the compiler options, so this should
end up in a state that seem users familiar with SwiftPM/Xcode will
expect.

Remove the private feature that was supporting this by honoring the
objc flag.

PiperOrigin-RevId: 558766347