Add macros support for Linux
This uses Swift's builtin feature detection to check what flags are
available. This probably means previous toolchains where macros weren't
complete would return true for this, but that's probably fine.

This also requires absolutizing TMPDIR since that is validated in a case
on Linux and errors otherwise. Hoping this doesn't cause issues.