Disable AEGs inside xcode_swift_toolchain
Because of the complexity, I would like to disable AEGs in this rule and proceed with the rest of the migration and flip the flag. I'll come back in the near future to enable them and add `toolchain` param to the affected actions.

PiperOrigin-RevId: 567319616
(cherry picked from commit ee7f493a87e206db35f724601c52d1a18d64eb50)