Remove the `cxx_interop` attribute and related transitions.
This approach using transitions to set the C++ interop flag isn't viable. Deleting it until I can rethink it; the major challenge is, in an explicit modules world, how to handle the fact that system frameworks should be supported in both modes without creating unique targets for the "C++" and "non-C++" versions of those targets.

PiperOrigin-RevId: 571086586