Update `swift_import` docs and add warning
Updates the `swift_import` documentation to provide more information on the `swiftinterface` attribute.
This attribute is required when importing pre-built Swift targets across Xcode and/or OS versions.

Additionally, adds a warning when `swiftinterface` is used alongside `swiftmodule` as this used to just drop `swiftinterface` silently.
We'll make this a failure in a new major version to not break users.

Co-authored-by: John Szumski <784312+jszumski@users.noreply.github.com>