Add the feature `swift.werror.<diagnostic-name>` to selectively allow certain warnings from the Swift compiler to be treated as errors.
The feature can be specified multiple times, with different diagnostic names.

The prefix `swift.werror...` is meant to evoke the `-Werror` Clang flag; `swift.warning_as_error...` and other variations felt too verbose.

This works by using the `-debug-diagnostic-names` flag to print the symbolic name of the diagnostic at the end of the message, in square brackets. Then the worker process intercepts stderr and looks for diagnostics with matching names, and modifies the message and exit code if there's a match.

PiperOrigin-RevId: 632470858