Create new `SwiftCompile{Module,Codegen}` actions that split out module compilation from object file compilation (and support batching of codegen jobs).
This does not enable parallelized compilation yet; it is gated behind the `swift.compile_in_parallel`.

Some work still needs to be done to fully implement this. Namely, we need to ensure that macro expansions are generated by the right action, and we're currently restricting batched codegen to builds that don't require indexstore so that any tools currently using indexstore can be updated to merge multiple stores (there would be one per batch).

PiperOrigin-RevId: 659954905