Remove `async` from the `main` function for the time being
This is causing some of our tests to terminate abnormally with an error out of the Swift runtime: "freed pointer was not the last allocation". We're going to need `async` eventually to support swift-testing, but in the meantime, this fixes the broken projects while I investigate further.

PiperOrigin-RevId: 662923954
(cherry picked from commit 9fb1eaaf4118609e1a99d1cb381acbe864ce05a3)
Signed-off-by: Brentley Jones <github@brentleyjones.com>