Pull the XCTest runner(s) and test sharding/filtering logic out of the generator and just make them regular Swift classes in the observer module
There's no compelling reason to generate all of this code; the only thing that actually needs to be generated is the list of tests for Linux, and the small `main` that invokes it. For simplicity, we generate it on Darwin platforms as well, even though it's not strictly necessary because all the discovery there is dynamic.

This change also removes the detached `Task` hack that I introduced previously. This worked when all the tests pass (the common case), but if a test failed, XCTest complained that the `recordIssue` method wasn't being called on the main thread. I'll revisit this in a subsequent change to support swift-testing.

PiperOrigin-RevId: 664761141
(cherry picked from commit 6d6727ae0ad5805b74b9383d217230fd0c479100)
Signed-off-by: Brentley Jones <github@brentleyjones.com>