Improve error messaging around parallel compilation planning.
This gives users an actionable way to opt out if some future version of the Swift driver changes things and we need to adjust.

Also bring back the portion of the earlier rolled-back change that lets all codegen jobs finish before returning (to ensure all possible diagnostics get printed), and improve the layering check error message to include misspelling as one of the reasons it might occur.

PiperOrigin-RevId: 679207758