Start adding support for first-class mixed language Swift modules.
A mixed language module is one that contains C/Objective-C code as part of the same target as Swift code (unlike overlays, which are separate targets). Specifically, this allows Swift and C/Objective-C code in a module to mutually refer to each other; overlays only allow Swift code can refer to the underlying C/Objective-C but not the other way around. Another important distinction is that the Swift code in a mixed language module must be unconditionally compiled and linked into dependents' binaries, even if it is only used by C/Objective-C clients (because the C/Objective-C code in the module itself could refer to the Swift code).

This initial change stages in a simple case of adding C/Objective-C source files to a `swift_*` target. Later changes will add support for headers, which add more complexity since we need to perform additional Clang module compilation actions to support them.

PiperOrigin-RevId: 681564997