Use a new provider to wrap symbol graph extraction used for test discovery.
Since symbol graph extraction is an implementation detail of `swift_test`, it shouldn't squat on the provider exclusively. Any other aspect invocation trying to extract symbol graphs on the same target will run into the "provider was provided twice" error.

To remedy this, introduce the ability for someone creating their own symbol graph aspect to automatically wrap the `SwiftSymbolGraphInfo` provider in one of their own choosing, and to add a distinguishing string to the output directory name so that outputs likewise don't collide. (In the future, we may want to do this automatically by hashing the active aspect IDs, but at this time I prefer not to introduce that complexity here.)

PiperOrigin-RevId: 685710974