add `swift.suppress_warnings` feature (#1489)
This is something I could use in my codebase. I would like have a way to
propagate the `-suppress-warnings` flag to all swift libraries, and
specifically, selectively disable it in my REPO.bazel and individual
targets. This would allow me to globally suppress warnings for external
repositories, without enabling it in my own code, and without needing to
juggle the state of the flag setting (`-suppress-warnings` and other
diagnostic flags are mutually exclusive afaik). This is the same as how
we organize the `swift.warnings_as_errors` feature today.

Companion change in apple_support is bazelbuild/apple_support#367