Only disable `parse_headers` when calling `cc_common.compile` on a set of headers that include the Swift generated header.
Disabling `parse_headers` universally for any action in a Swift context is too aggressive, because it prevents those actions from running on legitimate headers used in, say, a mixed C/Swift library. It also impacts feature configurations created in aspects applied to C libraries.

Only the generated header is known to be non-compliant for this action, so we lazily create a second feature configuration that is only passed to that specific compilation. Even though this compilation may include the other public headers, we don't skip `parse_headers` on those because there would have been a _previous_ compilation earlier in the `swift_common.compile` to compile the set of headers _excluding_ the generated header, and that one will execute the `parse_headers` actions.

After additional testing, I may move `cc_include_scanning` into this lazy feature config, for the same reason.

PiperOrigin-RevId: 731304506