Implement `swift_synthesize_interface_aspect`.
This aspect uses the new `swift-synthesize-interface` driver mode that will be available in Swift 6.1 to generate a file that contains the synthesized Swift interface for a module (such as a C/Obj-C module), similar to the functionality provided by SourceKit.

This can be used by IDEs and other clients to generate those interfaces without the complexity of a SourceKit invocation, which often do not play nicely with advanced build scenarios.

PiperOrigin-RevId: 731345457