Introduce a pure Swift runfiles library (#1310)
Fixes #890

I followed guidance from @fmeum and based this implementation on the
runfiles library of `rules_python` where applicable since this was
pointed as the reference implementation.

In addition to `rules_python` implementation, this implementation uses a
similar mechanism as the C++ implementation for deducing the
`RUNFILES_DIR` and `RUNFILES_MANIFEST_FILE` location based on `argv0`.

---------

Co-authored-by: Brentley Jones <github@brentleyjones.com>
Co-authored-by: Luis Padron <heyluispadron@gmail.com>
Co-authored-by: Fabian Meumertzheim <fabian@meumertzhe.im>