Ensure that `swift_synthesize_interface_aspect` generates unique outputs.
The aspect generated a file named `<target>.synthesized.swift`, which was incorrect if a target had multiple modules in its `SwiftInfo.direct_modules` field. It would also be incorrect to just switch to `<module name>.synthesized.swift`, because two targets could potentially re-export the same module.

To fully disambiguate, we create an intermediate directory named for the target, and then use the module name as the filename in that directory.

PiperOrigin-RevId: 740414505