Add support for generating an indexstore from a `.swiftinterface` file.
Starting with Swift 6.1 (Xcode 16.3), the frontend correctly generates an indexstore from a `.swiftinterface` file when it is build with `-explicit-interface-module-build` (https://github.com/swiftlang/swift/pull/79485). Collect that indexstore if requested.

To do this, the worker has to find the `.swiftinterface` file and extract the flags from it in order to pass them to the frontend. For system frameworks, this is a bit more delicate, because we don't know the path to the actual `.swiftinterface` file at analysis time (it may be a _variant_ of the current architecture, like `arm64e` for `arm64`). So we add logic to the worker that infers the right file given the path to the parent `.swiftmodule` directory, using similar logic to the compiler itself.

PiperOrigin-RevId: 742225315