Fix the way we link to XCTest and Testing in Xcode.
The original technique we were using (weak linking and using `dlopen`) never actually worked the way it was intended, because we were still using compile-time-bound references to XCTest classes. If those weak references aren't bound when dyld loads the binary, they won't be when the library is later loaded by `dlopen`. The only reason this actually did work was because we were passing the rpaths to the actual Xcode developer dirs to the linker; and then, it only worked if your Xcode was at the same path (the problem we were actually trying to solve).

Instead, we should use the xcode-select symlinks as the rpaths. These will work no matter what the actual paths to Xcode are (as long as the version you've selected when you run the tests is the same as the one used to link the tests).

PiperOrigin-RevId: 752683881